Fix BasePlayer SuccessRate
Needed to cast at least one of the parameters to double to return double, otherwise returned 1 or 0.

+++ TexasHoldem/TexasHoldem.AI.TodorPlayer/PlayerStates/BasePlayerState.cs
@@ -17,7 +17,7 @@
 
         public string Name => this.GetType().Name;
 
-        public double SuccessRate => this.GamesWon / this.GamesPlayed;
+        public double SuccessRate => (double)this.GamesWon / this.GamesPlayed;
 
         public int GamesPlayed { get; set; }
 
@@ -44,6 +44,7 @@
         public virtual void StartRound(StartRoundContext context)
         {
             this.HandStrength = HandStrengthCalculator.Calculate(context.CommunityCards.ToList(), new List<Card> { this.FirstCard, this.SecondCard });
+
             this.CommunityCards = context.CommunityCards;
         }
